refactor: ChatDetail 비즈니스로직 => customHook 분리
