+++ client/src/Component/Organism/Chat/ChatListDetail.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { useRecoilValue } from "recoil";
 import { ChatInput } from "@Molecules/Chat/ChatInput";
 import { LinkButton } from "@Core/LinkButton";
-import { ChatDetail } from "./ChatDetail";
-import { ChatInputContainer } from "./ChatInputContainer";
 import { chatTarget } from "@Recoil/Atom";
+import { ChatDetail } from "./ChatDetail/ChatDetail";
+import { ChatInputContainer } from "./ChatInputContainer";
 
 export const ChatListDetail = () => {
   const { chatRoomId } = useRecoilValue(chatTarget);
