+++ client/src/Component/Template/ChatRoomLeft/ChatRoomLeft.tsx
@@ -1,5 +1,5 @@
 import { ChatInput } from "@Molecules/Chat/ChatInput";
-import { ChatDetail } from "@Organism/Chat/ChatDetail";
+import { ChatDetail } from "@Organism/Chat/ChatDetail/ChatDetail";
 
 export const ChatRoomLeft = () => {
   return (
